% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_digits.R
\name{pi_digits}
\alias{pi_digits}
\title{Compute Digits of Pi}
\usage{
pi_digits(n, type = c("char", "numeric", "mpfr"))
}
\arguments{
\item{n}{Integer. Number of digits of pi to return after the decimal point.}

\item{type}{Character. Output type: "char", "numeric", or "mpfr".}
}
\value{
A character string, numeric value, or mpfr object depending on `type`.
}
\description{
Compute the first n digits of pi using arbitrary precision arithmetic.
Results are memoised so repeated calls for the same n are instantaneous.
}
\examples{
pi_digits(10)
pi_digits(50)
pi_digits(20, type = "numeric")

}
